+++ CHANGELOG.md
@@ -38,6 +38,12 @@ Feature enhancements:
   Implement the Tilde operator (@letmutx)
 - [FEATURE #189](https://github.com/jasonwilliams/boa/pull/189):
   Implement Array.prototype.includes (incl tests) (@simonbrahan)
+- [FEATURE #180](https://github.com/jasonwilliams/boa/pull/180):
+  Implement Array.prototype.slice (@muskuloes @letmutx)
+- [FEATURE #152](https://github.com/jasonwilliams/boa/issues/152):
+  Short Function syntax (no arguments)
+- [FEATURE #164](https://github.com/jasonwilliams/boa/issues/164):
+  Implement Array.prototype.fill() (@bojan88)
 - Array tests: Tests implemented for shift, unshift and reverse, pop and push (@muskuloes)
 - Demo page has been improved, new font plus change on input. Thanks @WofWca